backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
redis==5.0.1
aioredis==2.0.1
anthropic==0.18.1
pytesseract==0.3.10
Pillow==10.2.0
opencv-python-headless==4.9.0.80
pyzbar==0.1.9
httpx==0.27.0
cryptography==42.0.2
python-dotenv==1.0.1
loguru==0.7.2
celery==5.3.6
flower==2.0.1
psycopg2-binary==2.9.9
greenlet==3.0.3
aiosqlite>=0.19.0
